requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==40.0.1
dj-database-url==0.5.0
Django==3.2.18
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.2.1
django-filter==21.1
django-health-check==3.17.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==0.10.0
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2021.3
requests==2.28.2
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.8
zope.interface==6.0
